+++ repository/knowledge_wikipedia.py
@@ -4,21 +4,18 @@ from __future__ import annotations
 from datetime import datetime
 from typing import Sequence
 
-
 import numpy as np
 from psycopg.rows import dict_row
 from psycopg import sql
-from psycopg.types.json import Json
 
 from repository.connection_pool import ConnectionPoolPG
 from repository.model import DocumentRecord, WikipediaDbRecord
 
 
 VECTOR_TABLE_NAME = "documents"
-RUN_HISTORY_TABLE_NAME = "run_history"
 
-class WikipediaPgRepository:
-    """Lightweight repository to write Wikipedia records into Postgres/pgvector."""
+class KnowledgeWikipedia:
+    """Repository to write Wikipedia records into Postgres/pgvector."""
 
     _pool: ConnectionPoolPG
     
@@ -27,8 +24,7 @@ class WikipediaPgRepository:
         pool: ConnectionPoolPG
     ) -> None:
         self._pool = pool
- 
-    # --- Embeddings Table (documents) ---- #
+
 
     def insert(self, row: WikipediaDbRecord) -> None:
         """Insert row"""
@@ -175,37 +171,4 @@ class WikipediaPgRepository:
 
         with self._pool.connection() as conn, conn.cursor() as cur:
             cur.execute(query_sql, (pid, source))
-            conn.commit()
-
-    # --- Run History Table (run_history) ---- #
-
-    def run_history_table_rows(self):
-        """Query all rows from the run_history table for debugging/observability purposes."""
-
-        query_sql = sql.SQL(
-                """
-                SELECT * FROM {table} ORDER BY id DESC;
-                """
-            ).format(table=sql.Identifier(RUN_HISTORY_TABLE_NAME))
-
-        with self._pool.connection() as conn, conn.cursor(row_factory=dict_row) as cur:
-            cur.execute(query_sql)
-            rows = cur.fetchall()
-        return rows
-
-    def insert_history_table_log(self, run_id: int, service_name: str, status: str, metadata: dict | None,
-                                 timestamp: datetime):
-        # pylint: disable=too-many-arguments
-        # pylint: disable=too-many-positional-arguments
-        """Insert a log entry into the history table"""
-
-        query_sql = sql.SQL(
-            """
-            INSERT INTO {table} (run_id, service_name, status, metadata, timestamp)
-            VALUES (%s, %s, %s, %s, %s)
-            """
-        ).format(table=sql.Identifier(RUN_HISTORY_TABLE_NAME))
-
-        with self._pool.connection() as conn, conn.cursor() as cur:
-            cur.execute(query_sql, (run_id, service_name, status, Json(metadata), timestamp))
             conn.commit()
